# doc/api.yml
openapi: 3.0.3
info:
  version: "latest"
  title: Prehakaton - API
  description: |
    XYXYXY
security:
  - BearerAuth: []
paths:
...
  "/courses":
    get:
      tags:
        - Courseing
      summary: Get all courses.
      operationId: getCourses
      parameters:
        - name: page
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Page"
        - name: page_size
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PageSize"
        - name: isSuggest
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of courses that meet the criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  "/courses/{pid}":
    get:
      tags:
        - Courseing
      summary: Get course by identifier.
      operationId: getCourseById
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The identified course.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
    put:
      tags:
        - Courseing
      summary: Crupdate course by identifier.
      operationId: crupdateCourseById
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Course to crupdate
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
      responses:
        "201":
          description: The crupdated course.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
    delete:
      tags:
        - Courseing
      summary: Delete course by identifier.
      operationId: deleteCourseById
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The deleted course.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  "/users/{uid}/courses/{xid}/react":
    put:
      tags:
        - Courseing
      summary: Reacte an courses.
      operationId: crupdateReactionCourses
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
        - name: xid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Reaction to crupdate
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reaction"
      responses:
        "200":
          description: List of courses that meet the criteria.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  "/courses/{xid}/react":
    get:
      tags:
        - Courseing
      summary: Get all ReactionCourses.
      operationId: getReactionCourses
      parameters:
        - name: xid
          in: path
          required: true
          schema:
            type: string
        - name: uid
          in: query
          required: false
          schema:
            type: string
        - name: have_like_reaction
          in: query
          required: false
          schema:
            type: boolean
        - name: have_vision
          in: query
          required: false
          schema:
            type: boolean
        - name: have_stars_number
          in: query
          required: false
          schema:
            type: boolean
        - name: have_comment
          in: query
          required: false
          schema:
            type: boolean
        - name: page
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Page"
        - name: page_size
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PageSize"
      responses:
        "200":
          description: List of courses that meet the criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
...
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BadRequestException"
    "403":
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotAuthorizedException"
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResourceNotFoundException"
    "429":
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TooManyRequestsException"
    "500":
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InternalServerException"
  schemas:
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
    Role:
      type: string
      enum:
        - CLIENT
        - MANAGER
    Sex:
      type: string
      enum:
        - M
        - F
        - OTHER
    ReactionType:
      type: string
      enum:
        - LIKE
        - DISLIKE
    UserStatus:
      type: string
      enum:
        - ENABLED
        - BANISHED
...
    StatusEnum:
      type: string
      enum:
        - planning
        - confirmed
        - in_progress
        - completed

        ## resource model
    ReactionsProperties:
      properties:
        like_number:
          type: integer
        view_number:
          type: integer
        star_medium:
          type: number
          format: float
    HaveImageProperties:
      properties:
        image_url:
          type: string
        picture_is_implemented:
          type: boolean
    Course:
      allOf:
        - $ref: "#/components/schemas/ReactionsProperties"
        - $ref: "#/components/schemas/HaveImageProperties"
      properties:
        id:
          type: string
        creation_datetime:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CourseCategory"
        followers:
          type: array
          items:
            $ref: "#/components/schemas/User"
        interested_users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/CourseSession"
        status:
          $ref: "#/components/schemas/StatusEnum"
        price:
          type: number
          format: float
...
    CourseSession:
      allOf:
        - $ref: "#/components/schemas/ReactionsProperties"
      properties:
        id:
          type: string
        creation_datetime:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        course:
          $ref: "#/components/schemas/Course"
        location:
          $ref: "#/components/schemas/Location"
        professor:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        start_datetime:
          type: string
          format: date-time
        end_datetime:
          type: string
          format: date-time
    Location:
      properties:
        id:
          type: string
        creation_datetime:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        name:
          type: string
        description:
          type: string
        latitude:
          type: string
        longitude:
          type: string
...
    Reaction:
      properties:
        id:
          type: string
        likeReaction:
          $ref: "#/components/schemas/ReactionType"
        creation_datetime:
          type: string
          format: date-time
        last_update_datetime:
          type: string
          format: date-time
        user_id:
          type: string
        subject_id:
          type: string
        vision:
          type: boolean
        StarsNumber:
          type: integer
        comment:
          type: string
    CreateUser:
      properties:
        id:
          type: string
        last_name:
          type: string
        first_name:
          type: string
        birth_date:
          type: string
          format: date
        email:
          type: string
        username:
          type: string
        status:
          $ref: "#/components/schemas/UserStatus"
        sex:
          $ref: "#/components/schemas/Sex"
        role:
          $ref: "#/components/schemas/Role"
    User:
      allOf:
        - $ref: "#/components/schemas/CreateUser"
      properties:
        photo_url:
          type: string
        profile_banner_url:
          type: string
        entrance_datetime:
          type: string
          format: date-time
...
    CourseCategory:
      type: object
      properties:
        name:
          type: string
          description: "Nom de la categorie"
        description:
          type: string
          description: "Détails supplémentaires concernant la categorie"

    ## exception model
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: InternalServerException
        message: Unexpected error